tools/commcare_hq_mcp/requirements.txt
mcp>=1.20.0
httpx>=0.27.0
python-dotenv>=1.0.0
